+++ test/utilities.py
@@ -1,5 +1,7 @@
 import numpy as np
 from pathlib import Path
+
+from numpy import random
 from pvinspect.data import (
     Image,
     ModuleImage,
@@ -8,6 +10,11 @@ from pvinspect.data import (
     EL_IMAGE,
 )
 from typing import List, Dict, Any, Optional
+import uuid
+
+
+def random_filename(ext: str = ".tif") -> str:
+    return uuid.uuid4().hex + ext
 
 
 def assert_equal(value, target, precision=1e-3):
@@ -25,7 +32,7 @@ def random_image(lazy: bool = False, **kwargs) -> Image:
     if "modality" not in kwargs.keys():
         kwargs["modality"] = EL_IMAGE
     if "path" not in kwargs.keys():
-        kwargs["path"] = Path() / "test.tif"
+        kwargs["path"] = Path() / random_filename()
 
     return Image(data, **kwargs)
 
@@ -41,7 +48,7 @@ def random_uint_image(lazy: bool = False, **kwargs) -> Image:
     if "modality" not in kwargs.keys():
         kwargs["modality"] = EL_IMAGE
     if "path" not in kwargs.keys():
-        kwargs["path"] = Path() / "test.tif"
+        kwargs["path"] = Path() / random_filename()
 
     return Image(data, **kwargs)
 
@@ -57,19 +64,28 @@ def random_ubyte_image(lazy: bool = False, **kwargs) -> Image:
     if "modality" not in kwargs.keys():
         kwargs["modality"] = EL_IMAGE
     if "path" not in kwargs.keys():
-        kwargs["path"] = Path() / "test.tif"
+        kwargs["path"] = Path() / random_filename()
 
     return Image(data, **kwargs)
 
 
 def random_module_image() -> ModuleImage:
     data = np.random.random((10, 10))
-    return ModuleImage(data, modality=EL_IMAGE, path=Path() / "test.tif")
+    return ModuleImage(data, modality=EL_IMAGE, path=Path() / random_filename())
 
 
 def random_image_sequence(
-    N: int = 3, meta: Optional[List[Dict[str, Any]]] = None, **kwargs
+    N: int = 3,
+    meta: Optional[List[Dict[str, Any]]] = None,
+    random_meta: bool = False,
+    **kwargs
 ) -> ImageSequence:
+    if random_meta and meta is None:
+        random_keys = [uuid.uuid4().hex for _ in range(10)]
+        meta = list()
+        for _ in range(N):
+            meta.append({k: random.randint(0, 10000) for k in random_keys})
+
     if meta is None:
         imgs = [random_image(**kwargs) for i in range(N)]
     else: