+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New
 
 * `save_images` now has an additional argument `filename_hook` that allows to modify the generation of image filenames
+* Meta data is now cached to speed up loading of long sequences
 
 ### Fixed